requirements.txt
langchain
openai
PyPDF2
python-dotenv
gradio
faiss-cpu
datetime
Chroma
